Add files via upload
Added an arbitrary optimal region.  If the Enter key is pressed, and there are no needs to improve rules, and there is no command entered, and the current state is not in the optimal region, a plan will be made to change the current state to be within the optimal region.