Add files via upload
Changed the description under the "Optimistic Union" section to only use two squares for unions.  Using more squares is possible, but calculating rule unions is harder and the risk of being wrong is greater.  Changed the display of a Region, so differences between the two states that make a region can be more visible. (01) = x, while (10) =X.  So looking at Xx10, you know that the states are 1010, 0110.  A rule displayed as XXXX->XXXx indicates (10)->(01) in the right-most bit position.  Other minor changes.