Add files via upload
Fixed a bug with two-result squares (the bane of my existence). If two squares have multiple results, and there is exactly one possible combination that works, combine them into a group/region.  Zero (obvious), or two possible combinations, indicate no combination should be made.  Apparently the two combination possibility can happen in the intersection of two groups/regions that have two results and look OK separately, so I added a fix in action::group_pair_intersection_needs, which ignores the situation instead of seeking a sample in the intersection.  A better fix would be to invalidate both groups, but preventing them from reestablishing on the next pass is a problem. A few other things in progress.