Add files via upload
Changed so running once runs a function and releases all memory at end.  The command to run multiple times is "cargo run <number>".  The "run" command still runs to the end of the current session.  Changed the optimal region, per domain, so it can be None.  Added the ToRegion need to move the current state to the optimal region. Improved domain::random_depth_first_backward_chaining.  When making plans for a lot of needs, of the same priority, it will try 3 at a time until at least one has a plan.