Add files via upload
Changed domain:any_mutually_exclusive_changes to work with SomeStep references, their was a bug.  Improved the display of domains, change domain when all needs are for a different domain.  Improved the display of needs that cannot be done.  Some needs can be generated with a "01" or "10" bit when only Xx/Xx is available, if starting with "00" or "11" the needs cannot be satisfied, or if starting with "01" or "10", a need of "00" or "11" cannot be done, which could be fixed with more granular action changes, but it is educational as is.