Add files via upload
Improved UI to allow commands at any user input opportunity.  Added the "dcs" command to Display The Current State, if a number of commands are used that do not automatically redisplay the current state, the current state can scroll off the screen.   Fixed a bug in the "dn" do need command, it was not processing the correct need.