Add files via upload
A state input from the command line can be specified in hexadecimal, like s0xabc.  A non-X containing region (same as before) can be specified as a state.  The program can be easily altered to allow a mask (SomeMask:_from_string) or bits (<SomeBits>:from_string) instance to be specified in hexadecimal.