Add files via upload
Improved domain:check_optimal to allow for intersections of optimal regions, and intersections of intersections.  Intersections and optimal regions will be randomly chosen.
If there is an intersection of all optimal regions, and it is chosen and moved to, boredom and region changes will stop, as there are no other regions to move to.