Add files via upload
Fixed a bug in plan length selection in domainstore:choose_need.
Added logic to domainstore:choose_need.  When a need plan goal is not in a negative select region, choose plans that go through the least amount of negative select regions.
The need-plan display will show a negative sum, at the end, to indicate traversing negative select regions, like ...P:0[3,2,3,6,3]/-2
When displaying domain 0, the command "to r1000", if it has developed the needed action rules, will force it into a negative status, and generate a FromSelectRegion need.
When displaying domain 0, the command "to r0", if it has developed the needed action rules, will force it into a conflicted status.