Add files via upload
Changed the region struct to allow more than 2 states to form a union.  Such a region is displayed with a trailing plus sign.  When adding a group with a region containing GT 1 state, the states will be displayed.  The gps command can be used to display the defining states of a group region.  Confirming a group with a region containing GT 2 states, will convert the region to two states.