Add files via upload
For testing, I want to be able to force a bits struct to have use more than one integer, even though the size of the integers used may change.  It turns out that the bits:new_from_string needed a big change to support the "+" character, to indicate a new integer.  So 1+011 converts to int[0]: 1, int[1]:3.