Add files via upload
Changed functions involved in generating a plan, to return a Result enum instead of an Option enum.  This allows error messages to propagate up the call stack (no surprise to most viewers).  This allows better info to be displayed when using the "to" UI command to learn why a need target cannot be reached, and debugging tests.